fix GIT_EXEC_PATH to point to bundle/git-core
This makes it be the same as in the linux runshell.

It may be that bundle/ used to have all the necessary binaries in it,
but that seems unlikely. I think that the breakage fixed by my previous
commit masked this other breakage, and anything git execs from git-core
(as opposed to being buitin) has never worked in the OSX dmg.